move the ATSU message into the types folder
