+++ fbw-common/src/systems/datalink/common/src/messages/NotamMessage.ts
@@ -2,7 +2,7 @@
 //  SPDX-License-Identifier: GPL-3.0
 
 import { AtsuMessage, AtsuMessageType, AtsuMessageSerializationFormat } from './AtsuMessage';
-import { AtsuTimestamp } from './AtsuTimestamp';
+import { AtsuTimestamp } from '../types/AtsuTimestamp';
 
 export class NotamMessage extends AtsuMessage {
     public Identifier: string = '';
