+++ fbw-common/src/systems/datalink/common/src/messages/index.ts
@@ -3,7 +3,6 @@
 
 export * from './AtisMessage';
 export * from './AtsuMessage';
-export * from './AtsuTimestamp';
 export * from './CpdlcMessage';
 export * from './CpdlcMessageElements';
 export * from './Conversion';
