+++ fbw-common/src/systems/datalink/common/src/types/index.ts
@@ -1,6 +1,7 @@
 //  Copyright (c) 2023 FlyByWire Simulations
 //  SPDX-License-Identifier: GPL-3.0
 
+export * from './AtsuTimestamp';
 export * from './Clock';
 export * from './DatalinkCommunicationSystems';
 export * from './DatalinkStatusCodes';
